+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,69 +1,16 @@
 package roomescape.repository;
 
 import java.util.List;
-import java.util.Map;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import roomescape.controller.dto.ReservationCreateRequest;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.util.CustomDateTimeFormatter;
 
-@Repository
-public class ReservationRepository {
+public interface ReservationRepository {
+    List<Reservation> findReservations();
 
-    private final String findAllReservationQuery = """
-            SELECT
-                r.id as reservation_id,
-                r.name,
-                r.date,
-                t.id as time_id,
-                t.start_at as time_value
-            FROM reservation as r
-            inner join reservation_time as t
-            on r.time_id = t.id""";
+    Reservation findReservationById(Long id);
 
-    private final RowMapper<Reservation> reservationRowMapper = (rs, rowNum) ->
-            new Reservation(
-                    rs.getLong("id"),
-                    rs.getString("name"),
-                    CustomDateTimeFormatter.getLocalDate(rs.getString("date")),
-                    new ReservationTime(
-                            rs.getLong("time_id"),
-                            CustomDateTimeFormatter.getLocalTime(rs.getString("time_value"))
-                    )
-            );
+    Long createReservation(ReservationCreateRequest reservationCreateRequest);
 
-    private final JdbcTemplate jdbcTemplate;
+    void deleteReservationById(Long id);
 
-    public ReservationRepository(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public List<Reservation> findReservations() {
-        return jdbcTemplate.query(findAllReservationQuery, reservationRowMapper);
-    }
-
-    public Reservation findReservationById(Long createdReservationId) {
-        return jdbcTemplate.queryForObject(findAllReservationQuery + " WHERE r.id = ?",
-                reservationRowMapper,
-                createdReservationId);
-    }
-
-    public Long createReservation(ReservationCreateRequest reservationCreateRequest) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
-                .withTableName("reservation")
-                .usingGeneratedKeyColumns("id");
-        return simpleJdbcInsert.executeAndReturnKey(Map.of(
-                "name", reservationCreateRequest.name(),
-                "date", reservationCreateRequest.date(),
-                "time_id", reservationCreateRequest.timeId()
-        )).longValue();
-    }
-
-    public void deleteReservationById(Long id) {
-        jdbcTemplate.update("DELETE FROM reservation WHERE id = ?", id);
-    }
 }
